requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# HTTP Requests
requests

# TTS Providers
gtts>=2.5.0
# pyttsx3>=2.90  # Alternative

# Environment Variables
python-dotenv>=1.0.0